requirements.txt
# ── Core Framework ────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── ML / NLP ─────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4

# ── Document Parsing ─────────────────────────────────────────
pdfplumber>=0.10.0
python-docx>=1.1.0

# ── Data / Utils ─────────────────────────────────────────────
numpy>=1.24.0
python-multipart>=0.0.6

# ── Database (optional, for production) ──────────────────────
# asyncpg>=0.29.0
# sqlalchemy[asyncio]>=2.0.0
# neo4j>=5.15.0
# redis>=5.0.0

# ── Testing ──────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
